+++ sample/src/commonTest/kotlin/SuspendingFunctionTest.kt
@@ -1,6 +1,5 @@
-import tz.co.asoft.asyncTest
-import tz.co.asoft.delayForFun
-import kotlin.js.JsName
+import test.asyncTest
+import test.delayForFun
 import kotlin.test.Test
 import kotlin.test.assertEquals