+++ test-coroutines/src/commonMain/kotlin/test/AsyncTest.kt
@@ -1,4 +1,4 @@
-package tz.co.asoft
+package test
 
 import kotlinx.coroutines.CoroutineScope